similarity_service/requirements.txt
fastapi
uvicorn
torch
transformers
wandb
python-dotenv
pytest
httpx
